Cleaned up wrt code analysis
git-svn-id: https://svn.apache.org/repos/asf/maven/shared/trunk@1631511 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/maven/shared/utils/reflection/ReflectorTestHelper.java
@@ -43,7 +43,7 @@ class ReflectorTestHelper
 
     protected ReflectorTestHelper( Integer throwCount )
     {
-        if ( throwCount != null && throwCount.intValue() > 0 )
+        if ( throwCount != null && throwCount > 0 )
         {
             throw new HelperException( "Something" );
         }
@@ -73,7 +73,7 @@ class ReflectorTestHelper
 
     protected static ReflectorTestHelper getInstance( Integer throwCount )
     {
-        if ( throwCount != null && throwCount.intValue() > 0 )
+        if ( throwCount != null && throwCount > 0 )
         {
             throw new HelperException( "Something" );
         }
